Update for unit testing

+++ AntiVirus/SAV_SimpleAntiVirus/SimpleAntivirus/GUI/Views/Pages/IntegrityPage.xaml.cs
@@ -4,6 +4,7 @@ using SimpleAntivirus.ViewModels.Pages;
 using System.Windows.Controls;
 using System.Windows.Media;
 using System.Windows.Media.Animation;
+using SimpleAntivirus.AntiTampering;
 namespace SimpleAntivirus.GUI.Views.Pages
 {
     /// <summary>
@@ -53,9 +54,17 @@ namespace SimpleAntivirus.GUI.Views.Pages
         {
             EnableButton(false);
             // 0 > no selection, 1> no item found, 2 -> item deleted
-            int resultInt = ViewModel.DeleteItem();
-            DisplayResultDelete(resultInt);
-            ViewModel.PathSelected = null;
+
+            System.Windows.MessageBoxResult choice = System.Windows.MessageBox.Show("Are you sure you want to delete the selected integrity entries?", "Simple Antivirus", System.Windows.MessageBoxButton.OKCancel, MessageBoxImage.Warning);
+            if (choice == System.Windows.MessageBoxResult.OK)
+            {
+                int resultInt = ViewModel.DeleteItem();
+                DisplayResultDelete(resultInt);
+                ViewModel.PathSelected = null;
+                EnableButton(true);
+                return;
+            }
+            DisplayResultDelete(4);
             EnableButton(true);
         }
 
@@ -65,11 +74,11 @@ namespace SimpleAntivirus.GUI.Views.Pages
             UpdateEntries();
             if (success)
             {
-                MessageBox.Show("Data added to database", "Integrity Add Status", MessageBoxButton.OK, MessageBoxImage.Information);
+                MessageBox.Show("Integrity Entry Added: File(s) successfully added to integrity checking list.", "Simple Antivirus", MessageBoxButton.OK, MessageBoxImage.Information);
             }
             else
             {
-                MessageBox.Show("Data failed to be added to database\n Most likely cause is that a file was protected / in use, preventing the program from reading its contents", "Integrity Add Failure", MessageBoxButton.OK, MessageBoxImage.Error);
+                MessageBox.Show("Integrity Add Failure: File(s) failed to be added.\n Most likely cause is that a file was protected / in use, preventing the program from reading its contents", "Simple Antivirus", MessageBoxButton.OK, MessageBoxImage.Error);
             }
         }
 
@@ -82,16 +91,19 @@ namespace SimpleAntivirus.GUI.Views.Pages
             switch (returnId)
             {
                 case 0:
-                    MessageBox.Show("No Item Selected", "Selection Error", MessageBoxButton.OK, MessageBoxImage.Error);
+                    MessageBox.Show("Integrity Entry Delete Failed: No Item Selected", "Simple Antivirus", MessageBoxButton.OK, MessageBoxImage.Error);
                     break;
                 case 1:
-                    MessageBox.Show("Data could not be found in Database", "Incorrect Entry", MessageBoxButton.OK, MessageBoxImage.Error);
+                    MessageBox.Show("Integrity Entry Delete Failed: Entry could not be found in database", "Simple Antivirus", MessageBoxButton.OK, MessageBoxImage.Error);
                     break;
                 case 2:
-                    MessageBox.Show("Data removed successfully", "Data Item Deleted", MessageBoxButton.OK, MessageBoxImage.Information);
+                    MessageBox.Show("Integrity Entries removed successfully!", "Simple Antivirus", MessageBoxButton.OK, MessageBoxImage.Information);
                     break;
                 case 3:
-                    MessageBox.Show("Some data was removed successfully", "Some Data Item Deleted", MessageBoxButton.OK, MessageBoxImage.Information);
+                    MessageBox.Show("Integrity Entries Partially Deleted: Some entries were removed successfully.", "Simple Antivirus", MessageBoxButton.OK, MessageBoxImage.Information);
+                    break;
+                case 4:
+                    System.Windows.MessageBox.Show("Delete operation cancelled. The selected files will continue to be checked for integrity.", "Simple Antivirus", System.Windows.MessageBoxButton.OK, MessageBoxImage.Stop);
                     break;
             }
         }
@@ -127,10 +139,15 @@ namespace SimpleAntivirus.GUI.Views.Pages
             OpenFileDialog fileDialog = new Microsoft.Win32.OpenFileDialog();
             fileDialog.ShowDialog();
             string fileGet = fileDialog.FileName;
-            if (fileGet != "")
+            // Start AntiTampering Implementation
+            if (InputValSan.FilePathCharLimit(fileGet) && InputValSan.FilePathValidation(fileGet))
             {
-                result = await ViewModel.AddIntegrityPath(fileGet);
-                DisplayResultOfAdded(result);
+                fileGet = InputValSan.FilePathSanitisation(fileGet);
+                if (fileGet != "")
+                {
+                    result = await ViewModel.AddIntegrityPath(fileGet);
+                    DisplayResultOfAdded(result);
+                }
             }
             EnableButton(true);
         }
@@ -145,14 +162,19 @@ namespace SimpleAntivirus.GUI.Views.Pages
             OpenFolderDialog folderDialog = new Microsoft.Win32.OpenFolderDialog();
             folderDialog.ShowDialog();
             string folderGet = folderDialog.FolderName;
-            // Start load bar
-            // Send to view model the path of folder.
-            if (folderGet != "")
+            // Start AntiTampering Implementation
+            if (InputValSan.FilePathCharLimit(folderGet) && InputValSan.FilePathValidation(folderGet))
             {
-                DisplayLoading(true);
-                result = await ViewModel.AddIntegrityPath(folderGet);
-                DisplayLoading(false);
-                DisplayResultOfAdded(result);
+                folderGet = InputValSan.FilePathSanitisation(folderGet);
+                // Start load bar
+                // Send to view model the path of folder.
+                if (folderGet != "")
+                {
+                    DisplayLoading(true);
+                    result = await ViewModel.AddIntegrityPath(folderGet);
+                    DisplayLoading(false);
+                    DisplayResultOfAdded(result);
+                }
             }
             EnableButton(true);
         }
@@ -236,6 +258,18 @@ namespace SimpleAntivirus.GUI.Views.Pages
         private void See_Results_Click(object sender, RoutedEventArgs e)
         {
             NavigationServiceIntermediary.NavigationService.Navigate(typeof(IntegrityResultsPage));
+        }
+
+        private void CardExpander_Expanded(object sender, RoutedEventArgs e)
+        {
+            if (IntegrityConfigExpansion.IsExpanded)
+            {
+                IntegrityPageWhole.Height = 984;
+            }
+            else
+            {
+                IntegrityPageWhole.Height = 500;
+            }
         }
     }
 }
